producer/requirements.txt
requests
google-api-python-client
python-dateutil
youtube-transcript-api
youtube-comment-downloader
yt-dlp